requirements.txt
streamlit
google-generativeai
pymongo
python-telegram-bot
pandas
plotly
certifi
flask